FORCEINLINE Math library. Removed some compile warnings. Created MainThreadJob (Main loop). Implemented movement input for DebugCamera (It's working now).

+++ ZooidEngine/ZooidEngine/ResourceManagers/ShaderManager.h
@@ -21,7 +21,7 @@ namespace ZE
 		ShaderChain* getShaderChain(int id);
 	protected:
 
-		UShort getShaderTypeByName(const char* resourceFilePath);
+		UInt8 getShaderTypeByName(const char* resourceFilePath);
 
 		static ShaderManager* s_instance;
